task-68692 added updated_on colomn for tables

+++ application/models/MyAppModel.php
@@ -72,6 +72,11 @@ class MyAppModel extends FatModel
             return false;
         }
 
+        if (!$this->updateModifiedTime()) {
+            $this->error = $this->getError();
+            return false;
+        }
+
         return true;
     }
 
@@ -92,10 +97,40 @@ class MyAppModel extends FatModel
         $this->objMainTableRecord->assignValues($arr, $handleDates, $mysql_date_format, $mysql_datetime_format, $execute_mysql_functions);
     }
 
+    public function updateModifiedTime()
+    {
+        if (!($this->mainTableRecordId > 0)) {
+            $this->error = 'ERR_INVALID_REQUEST_ID';
+            return false;
+        }
+
+        $this->assignValues(array(static::tblFld('updated_on') => date('Y-m-d H:i:s')));
+        $flds = $this->getFlds();
+        if (count($flds) == 0) {
+            return true;
+        }
+
+        if (!$this->objMainTableRecord->update(array('smt' => $this->mainTableIdField . ' = ?', 'vals' => array($this->mainTableRecordId)))) {
+            $this->error = $this->objMainTableRecord->getError();
+            return false;
+        }
+        return true;
+    }
+
     public function deleteRecord($deleteLangData = false)
     {
+        if (!($this->mainTableRecordId > 0)) {
+            $this->error = 'ERR_INVALID_REQUEST_ID';
+            return false;
+        }
+
         if (!FatApp::getDb()->deleteRecords($this->mainTableName, array('smt'=>$this->mainTableIdField . ' = ?', 'vals'=>array($this->mainTableRecordId)))) {
             $this->error = FatApp::getDb()->getError();
+            return false;
+        }
+        
+        if (!$this->updateModifiedTime()) {
+            $this->error = $this->getError();
             return false;
         }
 
@@ -104,10 +139,11 @@ class MyAppModel extends FatModel
         }
 
         $prefix = substr(static::DB_TBL_PREFIX, 0, -1);
-        if (!FatApp::getDb()->deleteRecords($this->mainTableName.'_lang', array('smt'=>$prefix . 'lang_' . static::DB_TBL_PREFIX . 'id' . ' = ?', 'vals'=>array($this->mainTableRecordId)))) {
+        if (!FatApp::getDb()->deleteRecords($this->mainTableName . '_lang', array('smt' => $prefix . 'lang_' . static::DB_TBL_PREFIX . 'id' . ' = ?', 'vals' => array($this->mainTableRecordId)))) {
             $this->error = FatApp::getDb()->getError();
             return false;
         }
+        
         return true;
     }
 
@@ -259,6 +295,11 @@ class MyAppModel extends FatModel
         return $this->objMainTableRecord->getFlds();
     }
 
+    public function unsetFld($key)
+    {
+        $this->objMainTableRecord->unsetFld($key);
+    }
+
     public function getFldValue($key)
     {
         return $this->objMainTableRecord->getFldValue($key);
@@ -289,6 +330,10 @@ class MyAppModel extends FatModel
             $this->error = $this->objMainTableRecord->getError();
         }
 
+        if (!$this->updateModifiedTime()) {
+            $this->error = $this->getError();
+        }
+
         return $result;
     }
 
@@ -309,21 +354,15 @@ class MyAppModel extends FatModel
             $this->error = 'ERR_INVALID_REQUEST_ID';
             return false;
         }
-
-        $db = FatApp::getDb();
-        if (! $db->updateFromArray(
-            static::DB_TBL,
-            array(
-            static::DB_TBL_PREFIX . 'active' => $v
-            ),
-            array(
-            'smt' => static::DB_TBL_PREFIX . 'id = ?',
-            'vals' => array(
-                        $this->mainTableRecordId
-            )
-            )
-        )) {
-            $this->error = $db->getError();
+        $data = array(
+            static::tblFld('updated_on') => date('Y-m-d H:i:s'),
+            //static::tblFld('active') => $v
+        );
+        $this->assignValues($data);
+        $this->setFldValue(static::tblFld('active'), $v);
+        
+        if (!$this->objMainTableRecord->update(array('smt' => $this->mainTableIdField . ' = ?', 'vals' => array($this->mainTableRecordId)))) {
+            $this->error = $this->objMainTableRecord->getError();
             return false;
         }
